Overall update
Added .changeDir() to return the direction of the last change — returns
-1 or 1. Useful if you want to do the counting yourself.
Added .setPos() to set the position manually. Useful if you’re keeping
track of a value somewhere else. For instance a list of positions in an
array.
Added .limit to limit the value returned by .pos() — this handles the
modulo for you. It always returns a positive value.